requirements.txt
# AnjayHD - Image Enhancement
# Dependencies

flask>=2.0.0
werkzeug>=2.0.0
pillow>=9.0.0
numpy>=1.20.0
opencv-python>=4.5.0
